requirements.txt
aiofiles==24.1.0
altair==5.3.0
aniso8601==9.0.1
annotated-types==0.7.0
anywidget==0.9.13
appdirs==1.4.4
asgiref==3.8.1
backcall==0.2.0
bcrypt==4.2.1
blessings==1.7
bqplot==0.12.43
cachelib==0.9.0
click-plugins==1.1.1
color-operations==0.1.5
colour==0.1.5
cssselect2==0.2.1
descartes==1.1.0
docopt==0.6.2
duckdb==1.0.0
earthengine-api==0.1.400
ee==0.2
eerepr==0.0.4
filelock==3.15.4
Flask-Caching==2.3.0
flask-restx==1.3.0
future==1.0.0
gdown==5.2.0
geemap==0.32.0
GeoAlchemy2==0.16.0
geocoder==1.38.1
gitdb==4.0.11
GitPython==3.1.43
google-api-python-client==2.154.0
google-auth==2.29.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.0
googlemaps==4.10.0
h3==3.7.7
hpack==4.0.0
htmltools==0.5.2
Hypercorn==0.17.3
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipympl==0.9.4
ipython==8.12.3
ipython-genutils==0.2.0
ipytree==0.2.2
ipyvue==1.11.1
ipyvuetify==1.9.4
ipywidgets==8.1.3
jupyter-leaflet==0.19.0
laspy==2.5.4
lazrs==0.6.0
leafmap==0.35.9
linkify-it-py==2.0.3
llvmlite==0.42.0
localtileserver==0.10.3
lonboard==0.9.3
maplibre==0.2.6
markdown-it-py==3.0.0
matplotlib-scalebar==0.8.1
mdit-py-plugins==0.4.1
mdurl==0.1.2
morecantile==5.3.0
multipledispatch==0.6.0
munkres==1.1.4
numexpr==2.10.1
oauthlib==3.2.2
overpass==0.7
palettable==3.3.3
paramiko==3.5.0
pipreqs==0.5.0
plotly==5.21.0
pmtiles==3.3.0
priority==2.0.0
prompt-toolkit==3.0.36
protobuf==4.25.3
psycopg2==2.9.10
psygnal==0.11.1
pyarrow==17.0.0
PyCRS==1.0.2
pydantic==2.8.2
pydantic_core==2.20.1
pyepsg==0.4.0
pylas==0.4.3
PyNaCl==1.5.0
pyperclip==1.8.2
pystac==1.10.1
pystac-client==0.8.2
python-box==7.1.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyWavelets==1.6.0
Quart==0.19.6
questionary==2.0.1
ratelim==0.1.6
requests-oauthlib==2.0.0
rich==13.7.1
rio-cogeo==5.3.3
rio-tiler==6.6.1
server-thread==0.2.0
setuptools==70.1.1
shiny==0.10.2
smmap==5.0.1
snuggs==1.4.7
SQLAlchemy==2.0.36
sshtunnel==0.4.0
starlette==0.37.2
streamlit==1.37.0
tenacity==8.2.3
toml==0.10.2
uc-micro-py==1.0.3
uvicorn==0.30.1
vtk==9.2.6
watchfiles==0.22.0
websockets==12.0
wheel==0.43.0
whitebox==2.3.4
whiteboxgui==2.3.0
widgetsnbextension==4.0.11
wsproto==1.2.0
yarg==0.1.9
zstandard==0.22.0
